+++ src/map/mod.rs
@@ -1,0 +1,1 @@
+pub mod map_err_str;